dolls1/backend/requirements.txt
django
django-environ
djangorestframework
django-cors-headers
psycopg[binary]
pandas
openpyxl
duckdb
pydantic
celery
redis
gunicorn
django-filter
drf-spectacular
numpy
statsmodels
scikit-learn
xgboost
